requirements.txt
flask==2.3.2
openai==1.14.3
python-dotenv==1.0.1
httpx==0.27.0
gunicorn==22.0.0 